requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
ckanapi==4.8
docopt==0.6.2
idna==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
simplejson==3.19.3
six==1.16.0
text-unidecode==1.3
urllib3==2.2.2
